+++ rollup.config.mjs
@@ -1,16 +1,17 @@
 /* eslint-disable */
 
+import terser from "@rollup/plugin-terser";
 import typescript from "@rollup/plugin-typescript";
-import { terser } from "rollup-plugin-terser";
 
-import pkg from "./package.json";
+import pkg from "./package.json" assert { type: "json" };
 
 const banner = `/*! kage.js v${pkg.version}
  *  Licensed under ${pkg.license}
  *  ${pkg.homepage}
  */`;
 
-export default {
+/** @type {import("rollup").RollupOptions} */
+const options = {
 	input: "src/browser.ts",
 	output: [
 		{
@@ -29,5 +30,15 @@ export default {
 			plugins: [terser()],
 		},
 	],
-	plugins: [typescript({ tsconfig: false, target: "es5" })],
+	plugins: [
+		typescript({
+			compilerOptions: {
+				module: "ESNext",
+				declaration: false,
+				outDir: undefined,
+			},
+		}),
+	],
 };
+
+export default options;